Refactor - create utils file Move GetBotID method to bot.go
